feat: move the Trigger button out of the submenu and show a count of triggers assigned to the Part
