+++ apps/app/src/react/components/inputs/TriggerBtn/style.scss
@@ -1,0 +1,6 @@
+.triggercount {
+	position: absolute;
+	bottom: -12px;
+	right: 0;
+	color: white;
+}
